requirements.txt
requests
nltk
python-dotenv
